fix: derive chat draft context without effect sync
- move draft serialization and storage helpers into a shared input draft module
- initialize pending draft refs from stored draft state instead of syncing them in useEffect
- keep chat mention and draft tests passing with the new draft initialization flow
