+++ apps/desktop/src/chat/components/session-provider.tsx
@@ -9,6 +9,7 @@ import {
   useState,
 } from "react";
 
+import { getDraftContextRefs } from "~/chat/components/input/draft";
 import { dedupeByKey, type ContextRef } from "~/chat/context/entities";
 import {
   type DisplayEntity,
@@ -64,7 +65,9 @@ export function ChatSession({
   const { user_id } = main.UI.useValues(main.STORE_ID);
 
   const [pendingManualRefs, setPendingManualRefs] = useState<ContextRef[]>([]);
-  const [pendingDraftRefs, setPendingDraftRefs] = useState<ContextRef[]>([]);
+  const [pendingDraftRefs, setPendingDraftRefs] = useState<ContextRef[]>(() =>
+    getDraftContextRefs(sessionId),
+  );
 
   const onAddContextEntity = useCallback((ref: ContextRef) => {
     setPendingManualRefs((prev) =>
